# target/generated-sources/openapi/api/openapi.yaml
openapi: 3.0.3
info:
  description: Data Platform User Context Service
  title: user-context-service
  version: v1
servers:
- url: /
...
paths:
...
  /usercontexts:
    get:
      description: User must have LIST_USER_CONTEXTS permission
      operationId: getAllUserContexts
      parameters:
      - description: What page to grab
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - description: Number of elements on page
        explode: true
        in: query
        name: pageSize
        required: false
        schema:
          default: 25
          type: integer
        style: form
      - description: Sort order direction
        explode: true
        in: query
        name: direction
        required: false
        schema:
          default: ASC
          enum:
          - ASC
          - DESC
          type: string
        style: form
      - description: Http cache control header
        explode: false
        in: header
        name: Cache-Control
        required: false
        schema:
          enum:
          - no-cache
          - must-revalidate
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContextPage'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: gets all userContexts
      tags:
      - UserContexts
      x-accepts: application/json
    post:
      description: User must have CREATE_USER_CONTEXTS permission
      operationId: createUserContext
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserContext'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: creates userContext
      tags:
      - UserContexts
      x-accepts: application/json
      x-content-type: application/json
  /usercontexts/batch:
    post:
      description: User must have CREATE_USER_CONTEXTS permission
      operationId: createUserContextBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserContextBatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContextPage'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: creates userContext
      tags:
      - UserContexts
      x-accepts: application/json
      x-content-type: application/json
  /usercontexts/{usercontextId}:
    delete:
      description: User must have DELETE_USER_CONTEXTS permission
      operationId: deleteUserContextByUserContextId
      parameters:
      - description: usercontextId
        explode: false
        in: path
        name: usercontextId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: deleted successfully
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: deletes userContext by usercontextId
      tags:
      - UserContexts
      x-accepts: application/json
    get:
      description: User must have LIST_USER_CONTEXTS permission
      operationId: getUserContextByUserContextId
      parameters:
      - description: usercontextId
        explode: false
        in: path
        name: usercontextId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: gets userContext by usercontextId
      tags:
      - UserContexts
      x-accepts: application/json
    put:
      description: User must have UPDATE_USER_CONTEXTS permission
      operationId: updateUserContext
      parameters:
      - description: usercontextId
        explode: false
        in: path
        name: usercontextId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserContext'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: updates userContext via upsert
      tags:
      - UserContexts
      x-accepts: application/json
      x-content-type: application/json
  /usercontexts/user/{userId}/context/{contextId}/role/{roleId}:
    post:
      description: User must have CREATE_USER_CONTEXTS permission
      operationId: addRoleToUserContext
      parameters:
      - description: userId
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      - description: contextId
        explode: false
        in: path
        name: contextId
        required: true
        schema:
          type: string
        style: simple
      - description: roleId
        explode: false
        in: path
        name: roleId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: header
        name: Cache-Control
        required: false
        schema:
          default: no-cache
          type: string
        style: simple
      - description: additionalPermissions comma delimited
        explode: true
        in: query
        name: additionalPermissions
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: gets userContext by userId and contextId
      tags:
      - UserContexts
      x-accepts: application/json
  /usercontexts/user/{userId}/context/{contextId}/permissions:
    post:
      operationId: addPermissionsToUserContext
      parameters:
      - description: userId
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      - description: contextId
        explode: false
        in: path
        name: contextId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: header
        name: Cache-Control
        required: false
        schema:
          default: no-cache
          type: string
        style: simple
      - description: additionalPermissions comma delimited
        explode: true
        in: query
        name: additionalPermissions
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
      summary: gets userContext by userId and contextId
      tags:
      - UserContexts
      x-accepts: application/json
...
components:
  schemas:
...
    UserContextPage:
      example:
        elapsed: elapsed
        total: 1
        pageSize: 6
        correlationId: correlationId
        page: 0
        content:
        - modifiedTime: modifiedTime
          createdBy: createdBy
          roleId: roleId
          additionalRoles:
          - additionalRoles
          - additionalRoles
          createdTime: createdTime
          contextId: contextId
          modifiedBy: modifiedBy
          id: id
          userId: userId
          additionalPermissions:
          - additionalPermissions
          - additionalPermissions
          enabled: true
          transactionId: transactionId
        - modifiedTime: modifiedTime
          createdBy: createdBy
          roleId: roleId
          additionalRoles:
          - additionalRoles
          - additionalRoles
          createdTime: createdTime
          contextId: contextId
          modifiedBy: modifiedBy
          id: id
          userId: userId
          additionalPermissions:
          - additionalPermissions
          - additionalPermissions
          enabled: true
          transactionId: transactionId
      properties:
        content:
          items:
            $ref: '#/components/schemas/UserContext'
          type: array
        elapsed:
          type: string
        correlationId:
          type: string
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      type: object
...
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http